move ElementParser into xml package
